Features and Bug Fixes Added
NOTE: The content of this commit is authored by Cole Pendergraft.  Due to issues adding their user to the repository, they are unable to commit to the project themselves.

Program now can fetch data on number of contributions within a given timeframe, and can display analytical data on information collected.